requirements.txt
aiohttp==3.4.4
async-timeout==3.0.1
attrs==18.2.0
beautifulsoup4==4.6.3
bs4==0.0.1
certifi==2018.11.29
chardet==3.0.4
Click==7.0
crawler==0.0.2
decorator==4.3.0
defusedxml==0.5.0
dict2xml==1.6
Django==2.1.3
django-tastypie==0.14.2
Flask==1.0.2
Flask-SQLAlchemy==2.3.2
html5lib==1.0.1
idna==2.7
itsdangerous==1.1.0
Jinja2==2.10
lxml==4.2.5
MarkupSafe==1.1.0
multidict==4.5.2
public==2018.11.20
python-dateutil==2.7.5
python-mimeparse==1.6.0
pytz==2018.7
requests==2.20.1
self==2018.11.20
six==1.11.0
SQLAlchemy==1.2.14
urllib3==1.24.1
webencodings==0.5.1
Werkzeug==0.14.1
xmltodict==0.11.0
yarl==1.2.6
